requirements.txt
aiohttp==1.0.5
astroid==1.5.3
async-timeout==1.2.1
asyncio==3.4.3
beautifultable==0.3.0
cffi==1.10.0
chardet==3.0.3
cycler==0.10.0
discord.py==0.16.8
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
mongoengine==0.13.0
multidict==2.1.6
numpy==1.12.1
pycparser==2.17
pylint==1.7.4
pymongo==3.5.1
PyNaCl==1.0.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
tabulate==0.7.7
websockets==3.3
wrapt==1.10.11
